chore: remove build output logs for cleaner repository structure
