+++ src/commonMain/kotlin/ai/solace/zlib/bitwise/ArithmeticBitwiseOps.kt
@@ -1,3 +1,4 @@
+@file:Suppress("unused", "UNUSED_PARAMETER")
 package ai.solace.zlib.bitwise
 
 import ai.solace.zlib.common.ZlibLogger
