+++ src/commonMain/kotlin/ai/solace/zlib/bitwise/BitShiftEngine.kt
@@ -1,3 +1,4 @@
+@file:Suppress("unused", "UNUSED_PARAMETER")
 package ai.solace.zlib.bitwise
 
 /**
